+++ packages/blocks/blocksAntd/src/blocks/Card/Card.js
@@ -16,18 +16,12 @@
 
 import React from 'react';
 import { Card } from 'antd';
-import { blockDefaultProps, RenderHtml } from '@lowdefy/block-tools';
+import { blockDefaultProps, renderHtml } from '@lowdefy/block-tools';
 
 const CardBlock = ({ blockId, content, properties, methods, events }) => (
   <Card
     id={blockId}
-    title={
-      content.title ? (
-        content.title()
-      ) : properties.title ? (
-        <RenderHtml html={properties.title} methods={methods} />
-      ) : undefined
-    }
+    title={content.title ? content.title() : renderHtml({ html: properties.title, methods })}
     headStyle={methods.makeCssClass(properties.headerStyle, { styleObjectOnly: true })}
     bodyStyle={methods.makeCssClass(properties.bodyStyle, { styleObjectOnly: true })}
     bordered={properties.bordered}